Refactor Equiformer stacks onto shared so3/e3nn_nn primitives.
Move layout-specific ops out of architecture packages so each Equiformer
generation stays core+interaction, and place radial MLP / drop / SH readout
in blocks or equivariant modules instead of layers.
