+++ enerzyme/models/blocks/drop.py
@@ -1,0 +1,49 @@
+"""Representation-agnostic stochastic depth (shared residual-path dropout)."""
+
+from __future__ import annotations
+
+import torch
+import torch.nn as nn
+
+
+def drop_path(x, drop_prob: float = 0.0, training: bool = False):
+    """Drop paths (Stochastic Depth) per sample."""
+    if drop_prob == 0.0 or not training:
+        return x
+    keep_prob = 1 - drop_prob
+    shape = (x.shape[0],) + (1,) * (x.ndim - 1)
+    random_tensor = keep_prob + torch.rand(shape, dtype=x.dtype, device=x.device)
+    random_tensor.floor_()
+    return x.div(keep_prob) * random_tensor
+
+
+class DropPath(nn.Module):
+    """Drop paths (Stochastic Depth) per sample."""
+
+    def __init__(self, drop_prob=None):
+        super().__init__()
+        self.drop_prob = drop_prob
+
+    def forward(self, x):
+        return drop_path(x, self.drop_prob, self.training)
+
+    def extra_repr(self):
+        return "drop_prob={}".format(self.drop_prob)
+
+
+class GraphDropPath(nn.Module):
+    """Stochastic depth with per-graph masks via ``batch`` indices."""
+
+    def __init__(self, drop_prob=None):
+        super().__init__()
+        self.drop_prob = drop_prob
+
+    def forward(self, x, batch):
+        batch_size = batch.max() + 1
+        shape = (batch_size,) + (1,) * (x.ndim - 1)
+        ones = torch.ones(shape, dtype=x.dtype, device=x.device)
+        drop = drop_path(ones, self.drop_prob, self.training)
+        return x * drop[batch]
+
+    def extra_repr(self):
+        return "drop_prob={}".format(self.drop_prob)
