+++ enerzyme/models/layers/readout.py
@@ -2,7 +2,7 @@ from typing import Set, Literal, Dict, Optional, List, Callable
 from . import BaseFFLayer
 from ..blocks.mlp import DenseLayer, ResidualMLP, ResidualLayer
 from ..activation import ACTIVATION_KEY_TYPE, ACTIVATION_PARAM_TYPE, POSITIVE_ACTIVATION_KEY_TYPE, get_positive_activation_fn
-from ..irreps_tools import extract_scalar_0e
+from ..e3nn_nn import extract_scalar_0e
 import math
 import torch
 from torch.nn import Embedding, Identity, Linear, ModuleList, Module, Sequential
@@ -200,9 +200,8 @@ class BaseReadout(BaseFFLayer):
             # Final LinearRS is widened by shallow_ensemble_size; reshape happens
             # in split_readout_field_outputs via reshape_ensemble_head_output.
             from e3nn import o3
-            from ..equiformer.attention import _RESCALE
-            from ..equiformer.fast_activation import Activation
-            from ..equiformer.tensor_product import LinearRS
+            from ..equiformer.interaction import _RESCALE
+            from ..e3nn_nn import Activation, LinearRS
 
             ir_hid = o3.Irreps(f"{self.dim_feature_in}x0e")
             n_out = self.dim_feature_out * self.shallow_ensemble_size
@@ -453,253 +452,6 @@ class VelocityReadout(BaseReadout):
         return {
             self.ordered_output_fields[i]: output[:, i] for i in range(self.dim_feature_out)
         }
-
-
-class EquiformerGraphAttentionReadout(BaseFFLayer):
-    """Equivariant GraphAttention head mapping full irreps → atomic scalars.
-
-    Unlike :class:`SimpleReadout` (0e extract + MLP), this consumes the full
-    ``atom_feature`` irreps tensor and graph edges, producing one 0e channel
-    per ``output_fields`` entry (e.g. ``Ea``, ``Qa``). Requires
-    ``feature_irreps`` from a prior equivariant Core.
-
-    With ``shallow_ensemble_size > 1``, the final GraphAttention ``LinearRS``
-    proj is widened and fields keep a trailing ensemble axis.
-    """
-
-    def __init__(
-        self,
-        output_fields: Set[str],
-        built_layers: List[Module],
-        irreps_head: str = "16x0e+8x1o+4x2e",
-        num_heads: int = 2,
-        fc_neurons: Optional[List[int]] = None,
-        irreps_sh: str = "1x0e+1x1e+1x2e",
-        irreps_node_attr: str = "1x0e",
-        irreps_pre_attn: Optional[str] = None,
-        rescale_degree: bool = False,
-        nonlinear_message: bool = True,
-        alpha_drop: float = 0.0,
-        proj_drop: float = 0.0,
-        feature_irreps: Optional[str] = None,
-        num_rbf: Optional[int] = None,
-        shallow_ensemble_size: int = 1,
-        **kwargs,
-    ) -> None:
-        from e3nn import o3
-        from ..equiformer.attention import GraphAttention
-
-        ordered = sorted(list(output_fields))
-        super().__init__(
-            input_fields={
-                "atom_feature",
-                "idx_i_sr",
-                "idx_j_sr",
-                "vij_sr",
-                "rbf",
-                "batch_seg",
-            },
-            output_fields=set(ordered),
-        )
-        self.ordered_output_fields = ordered
-        self.dim_feature_out = len(ordered)
-        self.shallow_ensemble_size = int(shallow_ensemble_size)
-
-        irreps_str = (
-            feature_irreps
-            if feature_irreps is not None
-            else _resolve_feature_irreps(built_layers)
-        )
-        if irreps_str is None:
-            raise ValueError(
-                "EquiformerGraphAttentionReadout requires feature_irreps "
-                "(from an equivariant Core) or an explicit feature_irreps=..."
-            )
-        self.feature_irreps = o3.Irreps(irreps_str)
-        self.irreps_edge_attr = o3.Irreps(irreps_sh)
-        self.irreps_node_attr = o3.Irreps(irreps_node_attr)
-
-        if fc_neurons is None:
-            fc_neurons = [64, 64]
-        if num_rbf is None:
-            num_rbf = 16
-            for layer in reversed(built_layers):
-                if hasattr(layer, "num_rbf"):
-                    num_rbf = int(layer.num_rbf)
-                    break
-                if hasattr(layer, "fc_neurons") and isinstance(layer.fc_neurons, list) and layer.fc_neurons:
-                    num_rbf = int(layer.fc_neurons[0])
-                    break
-        self.fc_neurons = [num_rbf] + list(fc_neurons)
-
-        n_out = self.dim_feature_out * self.shallow_ensemble_size
-        self.head = GraphAttention(
-            irreps_node_input=self.feature_irreps,
-            irreps_node_attr=self.irreps_node_attr,
-            irreps_edge_attr=self.irreps_edge_attr,
-            irreps_node_output=o3.Irreps(f"{n_out}x0e"),
-            fc_neurons=self.fc_neurons,
-            irreps_head=o3.Irreps(irreps_head),
-            num_heads=num_heads,
-            irreps_pre_attn=irreps_pre_attn,
-            rescale_degree=rescale_degree,
-            nonlinear_message=nonlinear_message,
-            alpha_drop=alpha_drop,
-            proj_drop=proj_drop,
-        )
-
-    def get_output(
-        self,
-        atom_feature: Tensor,
-        idx_i_sr: Tensor,
-        idx_j_sr: Tensor,
-        vij_sr: Tensor,
-        rbf: Tensor,
-        batch_seg: Optional[Tensor] = None,
-    ) -> Dict[str, Tensor]:
-        from e3nn import o3
-
-        if atom_feature.ndim != 2:
-            raise ValueError(
-                "EquiformerGraphAttentionReadout expects 2D atom_feature "
-                f"(N, irreps.dim); got shape {tuple(atom_feature.shape)}"
-            )
-        if atom_feature.shape[-1] != self.feature_irreps.dim:
-            raise ValueError(
-                f"atom_feature last dim {atom_feature.shape[-1]} != "
-                f"feature_irreps.dim {self.feature_irreps.dim}"
-            )
-        n_atoms = atom_feature.shape[0]
-        if batch_seg is None:
-            batch_seg = torch.zeros(n_atoms, dtype=torch.long, device=atom_feature.device)
-        # Enerzyme edge convention: edge_src = idx_j, edge_dst = idx_i
-        edge_src = idx_j_sr
-        edge_dst = idx_i_sr
-        node_attr = torch.ones_like(atom_feature.narrow(1, 0, 1))
-        edge_sh = o3.spherical_harmonics(
-            l=self.irreps_edge_attr,
-            x=vij_sr,
-            normalize=True,
-            normalization="component",
-        )
-        outputs = self.head(
-            node_input=atom_feature,
-            node_attr=node_attr,
-            edge_src=edge_src,
-            edge_dst=edge_dst,
-            edge_attr=edge_sh,
-            edge_scalars=rbf,
-            batch=batch_seg,
-        )
-        expected_last = self.dim_feature_out * self.shallow_ensemble_size
-        if outputs.ndim != 2 or outputs.shape[-1] != expected_last:
-            raise ValueError(
-                f"GraphAttention output shape {tuple(outputs.shape)} != "
-                f"(N, {expected_last})"
-            )
-        return split_readout_field_outputs(
-            outputs,
-            self.ordered_output_fields,
-            self.shallow_ensemble_size,
-        )
-
-
-class EquiformerV2FeedForwardReadout(BaseFFLayer):
-    """Paper EquiformerV2 energy head: sphere FFN → per-atom scalars.
-
-    Consumes ``atom_sphere_feature`` and Core ``SO3_grid`` / FFN hyperparameters.
-    Default production stacks use ``SimpleReadout`` on ``atom_feature`` instead.
-
-    With ``shallow_ensemble_size > 1``, the final ``SO3_LinearV2`` is widened;
-    only l=0 channels are used as field × ensemble scalars.
-    """
-
-    def __init__(
-        self,
-        output_fields: Set[str],
-        built_layers: List[Module],
-        ffn_hidden_channels: Optional[int] = None,
-        keep_feature: bool = False,
-        shallow_ensemble_size: int = 1,
-        **_unused,
-    ) -> None:
-        out = set(output_fields)
-        if keep_feature:
-            out = out | {"atom_feature", "atom_sphere_feature"}
-        super().__init__(
-            input_fields={"atom_sphere_feature"},
-            output_fields=out,
-        )
-        self.ordered_output_fields = sorted(list(output_fields))
-        self.keep_feature = keep_feature
-        self.shallow_ensemble_size = int(shallow_ensemble_size)
-        self.dim_feature_out = len(self.ordered_output_fields)
-
-        core = None
-        for layer in reversed(built_layers):
-            if hasattr(layer, "SO3_grid") and hasattr(layer, "sphere_channels"):
-                core = layer
-                break
-        if core is None:
-            raise ValueError(
-                "EquiformerV2FeedForwardReadout requires a preceding EquiformerV2 Core"
-            )
-
-        from ..equiformer_v2.transformer_block import FeedForwardNetwork
-        from ..so3 import SO3_Embedding
-
-        self._SO3_Embedding = SO3_Embedding
-        n_out = self.dim_feature_out * self.shallow_ensemble_size
-        hidden = (
-            ffn_hidden_channels
-            if ffn_hidden_channels is not None
-            else core.ffn_hidden_channels
-        )
-        self.ffn = FeedForwardNetwork(
-            sphere_channels=core.sphere_channels,
-            hidden_channels=hidden,
-            output_channels=n_out,
-            lmax_list=list(core.lmax_list),
-            mmax_list=list(core.mmax_list),
-            SO3_grid=core.SO3_grid,
-            activation=getattr(core, "ffn_activation", "scaled_silu"),
-            use_gate_act=getattr(core, "use_gate_act", False),
-            use_grid_mlp=getattr(core, "use_grid_mlp", False),
-            use_sep_s2_act=getattr(core, "use_sep_s2_act", True),
-        )
-        self.lmax_list = list(core.lmax_list)
-        self.mmax_list = list(core.mmax_list)
-        self.sphere_channels = core.sphere_channels
-
-    def get_output(self, atom_sphere_feature: Tensor) -> Dict[str, Tensor]:
-        x = self._SO3_Embedding(
-            0,
-            self.lmax_list.copy(),
-            self.sphere_channels,
-            device=atom_sphere_feature.device,
-            dtype=atom_sphere_feature.dtype,
-        )
-        x.set_embedding(atom_sphere_feature)
-        # Node features after Core are full degree (mmax == lmax); match Core output.
-        x.set_lmax_mmax(self.lmax_list.copy(), self.lmax_list.copy())
-        node_out = self.ffn(x)
-        # l=0 channels only → (N, n_fields * ensemble)
-        scalars = node_out.embedding.narrow(1, 0, 1).squeeze(1)
-        result = split_readout_field_outputs(
-            scalars,
-            self.ordered_output_fields,
-            self.shallow_ensemble_size,
-        )
-        if self.keep_feature:
-            result["atom_sphere_feature"] = atom_sphere_feature
-            # Concatenate l=0,m=0 channels across resolutions (same as Core).
-            features = []
-            offset_res = 0
-            for i, lmax in enumerate(self.lmax_list):
-                features.append(atom_sphere_feature[:, offset_res, :])
-                offset_res = offset_res + int((lmax + 1) ** 2)
-            result["atom_feature"] = torch.cat(features, dim=-1)
-        return result
 
 
 class HirshfeldReadout(BaseFFLayer):
